api/requirements.txt
fastapi==0.109.*
uvicorn[standard]==0.27.*
google-cloud-bigquery==3.*
python-multipart==0.0.9
